PR Feedback - Separate Error Mapping
It looks like we should keep the mapping of hr results to strings. There are error.message results but those may contain paths. I'm still hesistent to hard code all of these but found nothing in the runtime that provides the specific string mapping from the  int hr result itself

https://github.com/dotnet/runtime/blob/f84f00920e1719071492b260375af0d4403d340e/src/libraries/Common/src/System/HResults.cs#L67 <- they have this but there is no reversal of their string onto the int as a mapping, and I don't want to take some weird dependency on this file
