+++ test/dotnetup.Tests/DnupE2Etest.cs
@@ -571,6 +571,4 @@ public class ReuseAndErrorTests
             output.Should().NotContain("Downloading", "Should not download when files already exist from SDK");
         }
     }
-
-    /// <summary>
 }
